moneymaker/requirements.txt
Flask==2.1.2
logfmter==0.0.6
pandas==1.4.2
prometheus_flask_exporter==0.20.3
python-dotenv==0.21.0
SQLAlchemy==1.4.27
polygon-api-client==1.6.2
openpyxl==3.0.10
psycopg2-binary==2.9.5
pandarallel==1.6.4

273.99